Implemented Facade Design Pattern and preparing for Abstract Factory Pattern.
