--- src/main/se/kth/iv1350/integration/IRegistry.java
@@ -1,9 +1,0 @@
-package se.kth.iv1350.integration;
-
-import se.kth.iv1350.model.Sale;
-
-public interface IRegistry<T, U> {
-    T getDataInfo(U dataID);
-    void updateRegister(Sale closedSale);
-
-}
