+++ src/main/se/kth/iv1350/integration/ItemRegistry.java
@@ -1,0 +1,9 @@
+package se.kth.iv1350.integration;
+
+import se.kth.iv1350.integration.dto.ItemDTO;
+import se.kth.iv1350.model.Sale;
+
+public interface ItemRegistry<T> {
+    ItemDTO getDataInfo(T dataID) throws ItemNotFoundInItemRegistryException;
+    void updateRegistry(Sale closedSale);
+}
