+++ src/main/se/kth/iv1350/startup/Main.java
@@ -3,6 +3,7 @@ import se.kth.iv1350.controller.Controller;
 import se.kth.iv1350.controller.OperationFailedException;
 import se.kth.iv1350.integration.RegisterCreator;
 import se.kth.iv1350.integration.RegistryHandler;
+import se.kth.iv1350.integration.TotalRevenueFileOutput;
 import se.kth.iv1350.view.View;
 import se.kth.iv1350.integration.Printer;
 
@@ -38,11 +39,8 @@ public class Main {
         try {
             Printer printer = new Printer();
             RegisterCreator registerCreator = new RegisterCreator();
-            Controller contr = new Controller(printer, registerCreator);
-            // TODO: should it not just be "created" in the Controller layer
-//            RegistryHandler registryHandler = RegistryHandler.getInstance();
-//            Controller contr = new Controller(printer, registryHandler);
-
+            Controller contr = new Controller(printer);
+            contr.addCashRegisterObserver(TotalRevenueFileOutput.getInstance());
             View view = new View(contr);
             view.basicFlow();
             view.AlternativeFlow();
